Remove all references to 'passable'
